[WebAuthn] Overrelease in error for multi-origin credentials rdar://123357813 https://bugs.webkit.org/show_bug.cgi?id=269942
Reviewed by Chris Dumez.

There's evidence for crashes being caused by the over-release of the dictionary inside
userInfo[NSUnderlyingErrorKey] for the error returned for multi-origin credentials. This
patch is to use RetainPtr in WebKit for this dictionary. I have not been able to reproduce
the crash, so this is speculative.

* Source/WebKit/UIProcess/WebAuthentication/Cocoa/WebAuthenticatorCoordinatorProxy.mm:
(WebKit::WebAuthenticatorCoordinatorProxy::performRequest):

Canonical link: https://commits.webkit.org/275466@main