[css-scroll-snap] Skip renderers with no element in updateSnapOffsetsForScrollableArea https://bugs.webkit.org/show_bug.cgi?id=263093 rdar://111579277
Reviewed by Simon Fraser.

We should skip renderers in updateSnapOffsetsForScrollableArea if they have no related element.
For generated content we do generate an associated element as well, so the only time a renderer's
element is null is for anonymous renderers (as specified in RenderElement.h as well), which we
shouldn't snap to.

* Source/WebCore/page/scrolling/ScrollSnapOffsetsInfo.cpp:
(WebCore::updateSnapOffsetsForScrollableArea):

Canonical link: https://commits.webkit.org/275523@main