[web-animations] `content-visibility` interpolation wrapper should be specific to that property https://bugs.webkit.org/show_bug.cgi?id=270439 rdar://124005154
Reviewed by NOBODY (OOPS!).

We introduced a new `NonNormalizedDiscretePropertyWrapper` class in 273742@main such that it may be
shared between `visibility` and `content-visibility`. But these two properties actually do not share
enough common traits to be dealt with a single class (see bug 270434). As such we do not need
`NonNormalizedDiscretePropertyWrapper` at all and instead introduce a dedicated wrapper for the
`content-visibility` property. We also add the `DECLARE_ALLOCATOR_WITH_HEAP_IDENTIFIER` and
`DEFINE_ALLOCATOR_WITH_HEAP_IDENTIFIER` calls to match other wrappers.

* Source/WebCore/animation/CSSPropertyAnimation.cpp:
(WebCore::CSSPropertyAnimationWrapperMap::CSSPropertyAnimationWrapperMap):