Check both browser and chrome objects in enumerateFramesAndNamespaceObjects https://bugs.webkit.org/show_bug.cgi?id=270657 rdar://123409359
Reviewed by Timothy Hatcher.

Some extensions use a polyfill to overwrite the browser object with their own proxy. The Blue Canoe extension was doing this,
and it led to WebExtensionContextProxy::enumerateFramesAndNamespaceObjects not being able to find the namespace object for the extension,
since we were only checking the `browser` object.

To fix this, check both `browser` and `chrome`, and use whichever one is valid.

* Source/WebKit/WebProcess/Extensions/WebExtensionContextProxy.cpp:
(WebKit::WebExtensionContextProxy::enumerateFramesAndNamespaceObjects):

Canonical link: https://commits.webkit.org/275809@main