Ensure that the SWServerWorker is in the correct state before finishing installation rdar://121429889
Reviewed by Youenn Fablet.

Ensure that the SWServerWorker is in the expected state and bail if this method is triggered on a SWServerWorker in a different state. This method is callable over CoreIPC passing a ServiceWorkerIdentifier. Passing the ID of a service worker in any other state will reach the RELEASE_ASSERT

* Source/WebCore/workers/service/server/SWServerWorker.cpp:
(WebCore::SWServerWorker::didFinishInstall):

Canonical link: https://commits.webkit.org/272448.701@safari-7618-branch