Fix rendering of ruby block with continuation children https://bugs.webkit.org/show_bug.cgi?id=270792
Reviewed by NOBODY (OOPS!).

RenderTreeBuilder::Ruby::findOrCreateParentForStyleBasedRubyChild() assumes that
if a Ruby block has an anonymous child, it must be a Ruby box. This assumption
is not correct, as when there is mixed block and inline children, there might
be anonymous blocks created for using with continuations.

This assumption is strictly enforced in Debug builds with an ASSERT, but
in Release builds it results in the wrong anonymous block being returned as
the candidate parent for the child to be inserted, which can cause
elements to be rendered in the wrong place.

* LayoutTests/fast/ruby/ruby-block-with-continuation-children.html: Added.
* LayoutTests/platform/glib/fast/ruby/ruby-block-with-continuation-children-expected.txt: Added.
* LayoutTests/platform/ios/fast/ruby/ruby-block-with-continuation-children-expected.txt: Added.
* LayoutTests/platform/mac/fast/ruby/ruby-block-with-continuation-children-expected.txt: Added.
* Source/WebCore/rendering/updating/RenderTreeBuilderRuby.cpp:
(WebCore::RenderTreeBuilder::Ruby::findOrCreateParentForStyleBasedRubyChild):