[Wasm-GC] Fix br_on_cast issue on BBQ https://bugs.webkit.org/show_bug.cgi?id=268848
Reviewed by Justin Michaud.

While this error occurs in sample code using br_on_cast, it appears the bug is
actually in struct.get. Instead of `loadIfNecessary`, an `allocate` was used
for the struct ref. This doesn't work in the case that the ref gets spilled to
the stack due to, e.g., a runtime call.

This is just easy to trigger with br_on_cast, which currently always does a
runtime call for the cast part on BBQJIT and will likely spill the ref.

* JSTests/wasm/gc/br_on_cast.js:
* Source/JavaScriptCore/wasm/WasmBBQJIT64.cpp:
(JSC::Wasm::BBQJITImpl::BBQJIT::addStructGet):
* Source/JavaScriptCore/wasm/WasmBBQJIT32_64.cpp:
(JSC::Wasm::BBQJITImpl::BBQJIT::addStructGet):

Canonical link: https://commits.webkit.org/275920@main