Save empty style sheet when saving web page resources https://bugs.webkit.org/show_bug.cgi?id=270711 rdar://121205393
Reviewed by Chris Dumez.

We currently replace URLs of empty style sheets with relative file paths when saving page resources, but we don't
actually create files for them (as the content is empty), so the saved web page will point to a file that does not
exist. To fix this, ensure we create an file on disk even the content is empty.

* Source/WebCore/loader/archive/cf/LegacyWebArchive.cpp:
(WebCore::addSubresourcesForCSSStyleSheetsIfNecessary):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/CreateWebArchive.mm:

Canonical link: https://commits.webkit.org/275935@main