Crash when quickly loading and unloading a Web Extension. https://webkit.org/b/270862 rdar://124126618
Reviewed by Jeff Miller.

When the extension controller is loading an extension context we move the local storage.
The lambda can escape and be called async, but was only capturing references. We need
to capture a protected this reference, and check if the context is still loaded
before proceeding after the move has finished.

* Source/WebKit/UIProcess/Extensions/Cocoa/WebExtensionContextCocoa.mm:
(WebKit::WebExtensionContext::load):

Canonical link: https://commits.webkit.org/275986@main