REGRESSION(275943@main) TestWebKitAPI tests timing out https://bugs.webkit.org/show_bug.cgi?id=270875 rdar://124475961
Reviewed by Ben Nham and Chris Dumez.

TestWebKitAPI, when running against iOS, is not a "real" UI application. As such, the UIWindows
it creates are not associated with any UIWindowScene. This breaks the foreground/background checking
logic in ApplicationStateTracker.

Update ApplicationStateTracker to treat the view as "not background" if it has a UIWindow but no
UIScene.

* Source/WebKit/UIProcess/ApplicationStateTracker.mm:
(WebKit::ApplicationStateTracker::setScene):

Canonical link: https://commits.webkit.org/276009@main